fix problem with ambiguous chords [bugfix]
The main regular expression was ambiguous for certain chords and it
was causing empty steno output in those cases. (I think it was causing
the python script to crash during the call to `re.fullmatch`.)

The ambiguity was coming from the fact that some letters appear
on both sides of the steno keyboard.

I fixed it by:

(1) ensuring there's always a dash between the right and left halves
of the chord, and

(2) modifying the regex to require that at least one of the modifier
keys (right F/P/L/T) to be pressed